+++ src/main/java/no/utleiesystem/bilutleie/repositories/UtleiekontorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import no.utleiesystem.bilutleie.entities.Utleiekontor;
 
-public interface UtleiekontorRepository extends CrudRepository<Utleiekontor, Integer>{
+public interface UtleiekontorRepository extends CrudRepository<Utleiekontor, Long>{
     
 }
